refactor: переписать логику работы контроллера в сервис
