+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,11 +1,13 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.exception.InvalidDurationException;
 import ru.yandex.practicum.filmorate.exception.InvalidFilm;
 import ru.yandex.practicum.filmorate.exception.InvalidReleaseDateException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.service.FilmService;
 
 import javax.validation.Valid;
 import java.time.Duration;
@@ -18,60 +20,28 @@ import java.util.List;
 @RestController
 @RequestMapping("/films")
 public class FilmController {
+    private final FilmService filmService;
+
+    @Autowired
+    public FilmController(FilmService filmService) {
+        this.filmService = filmService;
+    }
     private HashMap<Integer, Film> films = new HashMap<>();
     private int currentId = 1;
 
     @GetMapping()
     public List<Film> getAllFilms(){
-        log.info("Fetching all films");
-
-        return new ArrayList<>(films.values());
+        return filmService.getAllFilms();
     }
 
     @PostMapping
     public Film addFilm(@RequestBody @Valid Film film) {
-        validateReleaseDate(film);
-        validateDuration(film);
-
-        film.setId(currentId++);
-        log.info("Adding a new film: {}", film);
-
-        films.put(film.getId(), film);
-
-        return film;
+        return filmService.addFilm(film);
     }
 
     @PutMapping
     public Film updateFilm(@RequestBody @Valid Film film){
-        if(!films.containsKey(film.getId())){
-            log.warn("Unknown film: {}", film);
-            throw new InvalidFilm("Unknown film");
-        }
-        validateReleaseDate(film);
-        validateDuration(film);
-
-        log.info("Update film {} to film: {} ", film, films.get(film.getId()));
-
-        films.put(film.getId(), film);
-
-        return film;
+        return filmService.updateFilm(film);
     }
 
-    private void validateReleaseDate(Film film) {
-        LocalDate releaseDate = film.getReleaseDate();
-        LocalDate minReleaseDate = LocalDate.of(1895, 12, 28);
-        if (releaseDate.isBefore(minReleaseDate)) {
-            log.warn("Release date must not be earlier than 28 December 1895, but is: {}", releaseDate);
-            throw new InvalidReleaseDateException("Release date must not be earlier than 28 December 1895");
-        }
-    }
-
-    private void validateDuration(Film film){
-        Duration duration = film.getDurationValue();
-
-        if(duration.isNegative()){
-            log.warn("Duration cannot be negative.");
-            throw new InvalidDurationException("Duration cannot be negative.");
-        }
-    }
 }
